+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 0.8.0
-* Add `onAddedField` callback
+* Add `getDisposer` callback
 
 # 0.7.0
 * Pass `form` to validator function